Dockerfile: Update building trace tools slightly
We have not been picking a tag for the trace-cmd build process.
Currently the tip of libtraceevent fails to build. Address both problems
here by picking recent stable tags for libtraceevent and libtracefs
(trace-cmd has no recent tags). Further, as it is often reported that
this fails to build due to a race, stop using "make -j$(nproc)" as this
is also small enough of a set of builds to not be an issue.

Signed-off-by: Tom Rini <trini@konsulko.com>
(cherry picked from commit 4dfa4c14b8e5b37885f9eb281b3bc230938c5972)

[pickman] Conflict resolution: Applied -j removal to existing ci/master
trace-cmd pins; also affected other make -j commands in the file.