CI: Add "allyesconfig" to one of the build jobs
Now that we can have "make allyesconfig" build and link, add this type
of build to the job which builds host tools as well. In GitLab, make
this job rather than binman testsuite be the job which unblocks the next
stage of the pipeline. This is because we had been using that job for
"sandbox builds", and now that we have an explicit test for that, we
should use it.

Signed-off-by: Tom Rini <trini@konsulko.com>
(cherry picked from commit 1d59479362c7f6a681a2b3c14e97edff7c0f5a7c)

[fix-up: Resolved conflict in .gitlab-ci.yml - kept HEAD's extra rules]